LPS-179567 Extract CommercePriceListPriceModifiersScreenNavigationCategory methods to a CommercePriceListPriceModifiersScreenNavigationEntry component and extends CommercePriceListPriceModifiersScreenNavigationCategory in it
